* task-026: (127 commits)
  Rename IGraphicMemory.drawLine for drawSpriteRow
  Implement the rndVxByteTest
  Refactor rndVxByte to decouple random number gen
  Implement the jpV0AddrTest
  Implement the ldIAddrTest
  Implement the addIVxTest
  Implement the ldIVxTest
  Implement the ldIVxTest
  Sort tests and add stubs for the missing ones
  Simplify the 'ret' test
  Simplify the 'jpAddr' test
  Simplify the 'call' test
  Simplify the 'ldVxByte' test
  Simplify the 'skpVx' test
  Simplify the 'sknpVx' test
  Simplify the 'addVxVy' test
  Simplify the 'subVxVy' test
  Simplify the 'subnVxVy' test
  Fix assignation order on several instructions
  Simplify the 'sneVxVy' test
  ...

+++ src/chipeitlib/src/main/java/es/chipeit/lib/interfaces/IRegisters.kt
@@ -2,8 +2,8 @@ package es.chipeit.lib.interfaces
 
 internal interface IRegisters {
     val v: IMemory<Byte>
-    var i: Short
+    var i: Int
     var pc: Int
 
-    var sp: Byte
+    var sp: Int
 }
